fix: US-006 - Split llamaindex vectorstore.py into focused modules
Extract SearchOpsMixin (9 search methods) into search_ops.py,
GraphOpsMixin (velesql, explain, match_query) into graph_ops.py,
and filter logic into filter_ops.py.
vectorstore.py: 460 NLOC, search_ops.py: 311, graph_ops.py: 84,
filter_ops.py: 84. No assert statements. 72 tests pass.
