+++ integrations/llamaindex/src/llamaindex_velesdb/vectorstore.py
@@ -8,7 +8,6 @@ from __future__ import annotations
 
 import hashlib
 import logging
-import warnings
 from typing import Any, List, Optional
 
 from llama_index.core.schema import BaseNode, TextNode
@@ -33,23 +32,34 @@ from llamaindex_velesdb.security import (
     validate_weight,
     validate_sparse_vector,
 )
+from llamaindex_velesdb.filter_ops import metadata_filters_to_core_filter
+from llamaindex_velesdb.search_ops import SearchOpsMixin
+from llamaindex_velesdb.graph_ops import GraphOpsMixin
+
+# Re-export for backward compatibility and discoverability.
+__all__ = [
+    "VelesDBVectorStore",
+    "SearchOpsMixin",
+    "GraphOpsMixin",
+    "metadata_filters_to_core_filter",
+]
 
 logger = logging.getLogger(__name__)
 
 
 def _stable_hash_id(value: str) -> int:
     """Generate a stable numeric ID from a string using SHA256.
-    
+
     Python's hash() is non-deterministic across processes, so we use
     SHA256 for consistent IDs across runs.
 
     Uses 63 bits from SHA256 for a very low collision probability in
     real-world dataset sizes while keeping a positive integer compatible
     with VelesDB point IDs.
-    
+
     Args:
         value: String to hash.
-        
+
     Returns:
         Positive 63-bit integer ID compatible with VelesDB Core.
     """
@@ -58,7 +68,7 @@ def _stable_hash_id(value: str) -> int:
     return int.from_bytes(hash_bytes[:8], byteorder="big") & 0x7FFFFFFFFFFFFFFF
 
 
-class VelesDBVectorStore(BasePydanticVectorStore):
+class VelesDBVectorStore(SearchOpsMixin, GraphOpsMixin, BasePydanticVectorStore):
     """VelesDB vector store for LlamaIndex.
 
     A high-performance vector store backed by VelesDB, designed for
@@ -126,7 +136,7 @@ class VelesDBVectorStore(BasePydanticVectorStore):
                 - "sq8": 8-bit scalar quantization (4x memory reduction)
                 - "binary": 1-bit binary quantization (32x memory reduction)
             **kwargs: Additional arguments.
-            
+
         Raises:
             SecurityError: If any parameter fails validation.
         """
@@ -135,7 +145,7 @@ class VelesDBVectorStore(BasePydanticVectorStore):
         validated_collection = validate_collection_name(collection_name)
         validated_metric = validate_metric(metric)
         validated_storage_mode = validate_storage_mode(storage_mode)
-        
+
         super().__init__(
             path=validated_path,
             storage_mode=validated_storage_mode,
@@ -143,6 +153,10 @@ class VelesDBVectorStore(BasePydanticVectorStore):
             metric=validated_metric,
             **kwargs,
         )
+
+    # ------------------------------------------------------------------
+    # Static / class helpers
+    # ------------------------------------------------------------------
 
     @staticmethod
     def _metadata_from_payload(payload: dict) -> dict:
@@ -179,99 +193,14 @@ class VelesDBVectorStore(BasePydanticVectorStore):
 
         return VectorStoreQueryResult(nodes=nodes, similarities=similarities, ids=ids)
 
-    @staticmethod
-    def _normalize_filter_operator(raw_operator: Any) -> str:
-        """Normalize LlamaIndex filter operator to VelesDB Core condition type."""
-        if raw_operator is None:
-            return "eq"
-        if hasattr(raw_operator, "value"):
-            raw_operator = raw_operator.value
-
-        op = str(raw_operator).strip().lower()
-        op_map = {
-            "eq": "eq",
-            "==": "eq",
-            "=": "eq",
-            "neq": "neq",
-            "ne": "neq",
-            "!=": "neq",
-            "gt": "gt",
-            ">": "gt",
-            "gte": "gte",
-            ">=": "gte",
-            "lt": "lt",
-            "<": "lt",
-            "lte": "lte",
-            "<=": "lte",
-            "in": "in",
-            "contains": "contains",
-            "text_match": "contains",
-            "is_null": "is_null",
-            "null": "is_null",
-            "is_not_null": "is_not_null",
-            "not_null": "is_not_null",
-        }
-        if op in op_map:
-            return op_map[op]
-        raise ValueError(f"Unsupported metadata filter operator: {raw_operator}")
-
-    @classmethod
-    def _single_metadata_filter_to_condition(cls, metadata_filter: Any) -> dict:
-        """Convert one LlamaIndex MetadataFilter into VelesDB Core condition."""
-        field = getattr(metadata_filter, "key", None)
-        if not isinstance(field, str) or not field:
-            raise ValueError("Metadata filter key must be a non-empty string")
-
-        operator = cls._normalize_filter_operator(getattr(metadata_filter, "operator", None))
-        value = getattr(metadata_filter, "value", None)
-
-        if operator == "in":
-            if not isinstance(value, list):
-                raise ValueError("Metadata filter operator 'in' requires a list value")
-            return {"type": "in", "field": field, "values": value}
-        if operator == "contains":
-            if not isinstance(value, str):
-                raise ValueError("Metadata filter operator 'contains' requires a string value")
-            return {"type": "contains", "field": field, "value": value}
-        if operator in {"is_null", "is_not_null"}:
-            return {"type": operator, "field": field}
-
-        return {"type": operator, "field": field, "value": value}
-
     @classmethod
     def _metadata_filters_to_core_filter(cls, filters: Any) -> Optional[dict]:
         """Convert LlamaIndex MetadataFilters to VelesDB Core filter format."""
-        if filters is None:
-            return None
+        return metadata_filters_to_core_filter(filters)
 
-        if isinstance(filters, dict):
-            if "condition" in filters:
-                return filters
-            conditions = [{"type": "eq", "field": key, "value": value} for key, value in filters.items()]
-            if not conditions:
-                return None
-            if len(conditions) == 1:
-                return {"condition": conditions[0]}
-            return {"condition": {"type": "and", "conditions": conditions}}
-
-        raw_filters = getattr(filters, "filters", None)
-        if raw_filters is None:
-            return None
-
-        conditions = [cls._single_metadata_filter_to_condition(item) for item in raw_filters]
-        if not conditions:
-            return None
-        if len(conditions) == 1:
-            return {"condition": conditions[0]}
-
-        condition_mode = getattr(filters, "condition", None)
-        if hasattr(condition_mode, "value"):
-            condition_mode = condition_mode.value
-        mode = str(condition_mode).strip().lower() if condition_mode is not None else "and"
-        if mode not in {"and", "or"}:
-            raise ValueError(f"Unsupported metadata filter condition mode: {condition_mode}")
-
-        return {"condition": {"type": mode, "conditions": conditions}}
+    # ------------------------------------------------------------------
+    # Connection management
+    # ------------------------------------------------------------------
 
     def _get_db(self) -> velesdb.Database:
         """Get or create the database connection."""
@@ -281,13 +210,13 @@ class VelesDBVectorStore(BasePydanticVectorStore):
 
     def _get_collection(self, dimension: int) -> velesdb.Collection:
         """Get or create the collection.
-        
+
         Args:
             dimension: Expected vector dimension.
-            
+
         Returns:
             The VelesDB collection.
-            
+
         Raises:
             ValueError: If collection exists with different dimension.
         """
@@ -307,8 +236,8 @@ class VelesDBVectorStore(BasePydanticVectorStore):
                 existing_dim = info.get("dimension", 0)
                 if existing_dim != 0 and existing_dim != dimension:
                     raise ValueError(
-                        f"Collection '{self.collection_name}' exists with dimension {existing_dim}, "
-                        f"but got vectors of dimension {dimension}. "
+                        f"Collection '{self.collection_name}' exists with dimension "
+                        f"{existing_dim}, but got vectors of dimension {dimension}. "
                         f"Use a different collection name or matching dimension."
                     )
             self._dimension = dimension
@@ -318,6 +247,10 @@ class VelesDBVectorStore(BasePydanticVectorStore):
     def client(self) -> velesdb.Database:
         """Return the VelesDB client."""
         return self._get_db()
+
+    # ------------------------------------------------------------------
+    # Write operations
+    # ------------------------------------------------------------------
 
     def add(
         self,
@@ -332,7 +265,7 @@ class VelesDBVectorStore(BasePydanticVectorStore):
 
         Returns:
             List of node IDs that were added.
-            
+
         Raises:
             SecurityError: If parameters fail validation.
         """
@@ -365,7 +298,6 @@ class VelesDBVectorStore(BasePydanticVectorStore):
                     )
 
         collection = self._get_collection(dimension)
-
         points = []
         ids = []
 
@@ -377,28 +309,19 @@ class VelesDBVectorStore(BasePydanticVectorStore):
             node_id = node.node_id
             ids.append(node_id)
 
-            # Build payload
-            payload = {
-                "text": node.get_content(),
-                "node_id": node_id,
-            }
+            payload = {"text": node.get_content(), "node_id": node_id}
 
-            # Add metadata
             if hasattr(node, "metadata") and node.metadata:
                 for key, value in node.metadata.items():
                     if isinstance(value, (str, int, float, bool)):
                         payload[key] = value
 
-            # Convert node_id to int for VelesDB
-            int_id = _stable_hash_id(node_id)
-
             point = {
-                "id": int_id,
+                "id": _stable_hash_id(node_id),
                 "vector": embedding,
                 "payload": payload,
             }
 
-            # Attach sparse vector if provided (v1.5)
             if sparse_vectors is not None and idx < len(sparse_vectors):
                 point["sparse_vector"] = sparse_vectors[idx]
 
@@ -422,303 +345,18 @@ class VelesDBVectorStore(BasePydanticVectorStore):
         int_id = _stable_hash_id(ref_doc_id)
         self._collection.delete([int_id])
 
-    def query(
-        self,
-        query: VectorStoreQuery,
-        **kwargs: Any,
-    ) -> VectorStoreQueryResult:
-        """Query the vector store.
-
-        Args:
-            query: Vector store query with embedding and parameters.
-            **kwargs: Additional arguments.  Accepts ``sparse_vector`` (dict)
-                and ``sparse_index_name`` (str) for hybrid dense+sparse search
-                targeting a specific named sparse index.
-
-        Returns:
-            Query result with nodes and similarities.
-
-        Raises:
-            SecurityError: If parameters fail validation.
-        """
-        if query.query_embedding is None:
-            return VectorStoreQueryResult(nodes=[], similarities=[], ids=[])
-
-        dimension = len(query.query_embedding)
-        collection = self._get_collection(dimension)
-
-        k = query.similarity_top_k or 10
-
-        # Security: Validate k
-        validate_k(k)
-
-        # Extract sparse vector and optional index name from kwargs
-        sparse_vector = kwargs.get("sparse_vector")
-        if sparse_vector is not None:
-            validate_sparse_vector(sparse_vector)
-        sparse_index_name = kwargs.get("sparse_index_name")
-
-        core_filter = self._metadata_filters_to_core_filter(query.filters)
-
-        if sparse_vector is not None and core_filter is not None:
-            raise ValueError(
-                "sparse_vector cannot be combined with metadata filters. "
-                "Apply filters separately or omit the sparse_vector."
-            )
-
-        results = self._execute_query(
-            collection, query.query_embedding, k,
-            sparse_vector=sparse_vector,
-            sparse_index_name=sparse_index_name,
-            core_filter=core_filter,
-        )
-
-        return self._build_query_result(results)
-
-    def _execute_query(
-        self,
-        collection: Any,
-        query_embedding: List[float],
-        k: int,
-        *,
-        sparse_vector: Optional[dict] = None,
-        sparse_index_name: Optional[str] = None,
-        core_filter: Optional[dict] = None,
-    ) -> List[dict]:
-        """Execute the appropriate search variant on the collection.
-
-        Delegates to filtered, hybrid sparse, or plain dense search
-        depending on which optional arguments are provided.
-
-        Args:
-            collection: VelesDB collection object.
-            query_embedding: Dense query vector.
-            k: Number of results to return.
-            sparse_vector: Optional sparse vector dict for hybrid search.
-            sparse_index_name: Optional named sparse index to query.
-            core_filter: Optional metadata filter dict.
-
-        Returns:
-            List of search result dicts.
-        """
-        if core_filter is not None:
-            search_with_filter = getattr(collection, "search_with_filter", None)
-            if search_with_filter is None:
-                raise NotImplementedError(
-                    "Collection does not support 'search_with_filter' required for MetadataFilters."
-                )
-            return search_with_filter(query_embedding, top_k=k, filter=core_filter)
-
-        if sparse_vector is not None:
-            return self._run_sparse_search(
-                collection, query_embedding, sparse_vector, k,
-                sparse_index_name=sparse_index_name,
-            )
-
-        return collection.search(query_embedding, top_k=k)
-
-    def _run_sparse_search(
-        self,
-        collection: Any,
-        query_embedding: List[float],
-        sparse_vector: dict,
-        k: int,
-        *,
-        sparse_index_name: Optional[str] = None,
-    ) -> List[dict]:
-        """Run hybrid dense+sparse search, degrading to dense-only on failure.
-
-        Args:
-            collection: VelesDB collection object.
-            query_embedding: Dense query vector.
-            sparse_vector: Sparse vector dict mapping int term IDs to float weights.
-            k: Number of results to return.
-            sparse_index_name: Optional named sparse index to target.
-
-        Returns:
-            List of search result dicts.
-        """
-        search_kwargs: dict[str, Any] = {
-            "vector": query_embedding,
-            "sparse_vector": sparse_vector,
-            "top_k": k,
-        }
-        if sparse_index_name is not None:
-            search_kwargs["sparse_index_name"] = sparse_index_name
-
-        try:
-            return collection.search(**search_kwargs)
-        except RuntimeError:
-            warnings.warn(
-                "sparse_vector was provided but the collection does not have a sparse "
-                "index (no sparse vectors have been inserted). Falling back to "
-                "dense-only search. Insert points with sparse_vectors to enable "
-                "hybrid dense+sparse retrieval.",
-                UserWarning,
-                stacklevel=2,
-            )
-            return collection.search(query_embedding, top_k=k)
-
-    def query_with_score_threshold(
-        self,
-        query: VectorStoreQuery,
-        score_threshold: float = 0.0,
-        **kwargs: Any,
-    ) -> VectorStoreQueryResult:
-        """Query with similarity score threshold filtering.
-
-        This method enables similarity()-like filtering from VelesDB Core.
-        Only returns results with score >= threshold.
-
-        Args:
-            query: Vector store query with embedding and parameters.
-            score_threshold: Minimum similarity score (0.0-1.0 for cosine).
-                Only return nodes with score >= threshold.
-            **kwargs: Additional arguments.
-
-        Returns:
-            Query result with nodes above threshold.
-
-        Example:
-            >>> # Get only highly relevant results (>0.8 similarity)
-            >>> query = VectorStoreQuery(
-            ...     query_embedding=embedding,
-            ...     similarity_top_k=20
-            ... )
-            >>> result = vector_store.query_with_score_threshold(
-            ...     query, score_threshold=0.8
-            ... )
-        """
-        result = self.query(query, **kwargs)
-
-        if score_threshold > 0.0 and result.similarities:
-            filtered_indices = [
-                i for i, score in enumerate(result.similarities)
-                if score >= score_threshold
-            ]
-            return VectorStoreQueryResult(
-                nodes=[result.nodes[i] for i in filtered_indices] if result.nodes else [],
-                similarities=[result.similarities[i] for i in filtered_indices],
-                ids=[result.ids[i] for i in filtered_indices] if result.ids else [],
-            )
-
-        return result
-
-    def hybrid_query(
-        self,
-        query_str: str,
-        query_embedding: List[float],
-        similarity_top_k: int = 10,
-        vector_weight: float = 0.5,
-        **kwargs: Any,
-    ) -> VectorStoreQueryResult:
-        """Hybrid search combining vector similarity and BM25 text search.
-
-        Uses Reciprocal Rank Fusion (RRF) to combine results.
-
-        Args:
-            query_str: Text query for BM25 search.
-            query_embedding: Query embedding vector.
-            similarity_top_k: Number of results to return.
-            vector_weight: Weight for vector results (0.0-1.0). Defaults to 0.5.
-            **kwargs: Additional arguments.
-
-        Returns:
-            Query result with nodes and similarities.
-            
-        Raises:
-            SecurityError: If parameters fail validation.
-        """
-        # Security: Validate inputs
-        validate_text(query_str)
-        validate_k(similarity_top_k)
-        validate_weight(vector_weight, "vector_weight")
-        
-        dimension = len(query_embedding)
-        collection = self._get_collection(dimension)
-
-        results = collection.hybrid_search(
-            vector=query_embedding,
-            query=query_str,
-            top_k=similarity_top_k,
-            vector_weight=vector_weight,
-        )
-
-        return self._build_query_result(results)
-
-    def text_query(
-        self,
-        query_str: str,
-        similarity_top_k: int = 10,
-        **kwargs: Any,
-    ) -> VectorStoreQueryResult:
-        """Full-text search using BM25 ranking.
-
-        Args:
-            query_str: Text query string.
-            similarity_top_k: Number of results to return.
-            **kwargs: Additional arguments.
-
-        Returns:
-            Query result with nodes and similarities.
-            
-        Raises:
-            SecurityError: If parameters fail validation.
-        """
-        # Security: Validate inputs
-        validate_text(query_str)
-        validate_k(similarity_top_k)
-        
-        if self._collection is None:
-            return VectorStoreQueryResult(nodes=[], similarities=[], ids=[])
-
-        results = self._collection.text_search(query_str, top_k=similarity_top_k)
-
-        return self._build_query_result(results)
-
-    def batch_query(
-        self,
-        queries: List[VectorStoreQuery],
-        **kwargs: Any,
-    ) -> List[VectorStoreQueryResult]:
-        """Batch query with multiple embeddings in parallel.
-        
-        Raises:
-            SecurityError: If batch size exceeds limit.
-        """
-        if not queries:
-            return []
-        
-        # Security: Validate batch size
-        validate_batch_size(len(queries))
-
-        first_emb = queries[0].query_embedding
-        if first_emb is None:
-            return [VectorStoreQueryResult(nodes=[], similarities=[], ids=[]) 
-                    for _ in queries]
-
-        dimension = len(first_emb)
-        collection = self._get_collection(dimension)
-
-        searches = [{"vector": q.query_embedding, "top_k": q.similarity_top_k or 10}
-                    for q in queries if q.query_embedding is not None]
-
-        batch_results = collection.batch_search(searches)
-
-        return [self._build_query_result(res_list) for res_list in batch_results]
-
     def add_bulk(self, nodes: List[BaseNode], **add_kwargs: Any) -> List[str]:
         """Bulk insert optimized for large batches.
-        
+
         Raises:
             SecurityError: If batch size exceeds limit.
         """
         if not nodes:
             return []
-        
+
         # Security: Validate batch size
         validate_batch_size(len(nodes))
-        
+
         first_emb = nodes[0].get_embedding()
         if first_emb is None:
             raise ValueError("Nodes must have embeddings")
@@ -733,196 +371,14 @@ class VelesDBVectorStore(BasePydanticVectorStore):
             result_ids.append(nid)
             payload = {"text": node.get_content(), "node_id": nid}
             if hasattr(node, "metadata") and node.metadata:
-                payload.update({k: v for k, v in node.metadata.items() 
-                               if isinstance(v, (str, int, float, bool))})
+                payload.update({
+                    k: v for k, v in node.metadata.items()
+                    if isinstance(v, (str, int, float, bool))
+                })
             points.append({"id": _stable_hash_id(nid), "vector": emb, "payload": payload})
         if points:
             collection.upsert_bulk(points)
         return result_ids
-
-    def get_nodes(self, node_ids: List[str], **kwargs: Any) -> List[TextNode]:
-        """Retrieve nodes by their IDs."""
-        if not node_ids or self._collection is None:
-            return []
-        int_ids = [_stable_hash_id(nid) for nid in node_ids]
-        points = self._collection.get(int_ids)
-        result = []
-        for pt in points:
-            if pt:
-                p = pt.get("payload", {})
-                result.append(
-                    TextNode(
-                        text=p.get("text", ""),
-                        id_=p.get("node_id", ""),
-                        metadata=self._metadata_from_payload(p),
-                    )
-                )
-        return result
-
-    def get_collection_info(self) -> dict:
-        """Get collection configuration information."""
-        if self._collection is None:
-            return {"name": self.collection_name, "dimension": 0, "metric": self.metric, "point_count": 0}
-        return self._collection.info()
-
-    def flush(self) -> None:
-        """Flush all pending changes to disk."""
-        if self._collection is not None:
-            self._collection.flush()
-
-    def is_empty(self) -> bool:
-        """Check if the collection is empty."""
-        return self._collection is None or self._collection.is_empty()
-
-    def create_metadata_collection(self, name: str) -> None:
-        """Create a metadata-only collection (no vectors).
-
-        Useful for storing reference data that can be JOINed with
-        vector collections (VelesDB Premium feature).
-
-        Args:
-            name: Collection name.
-        """
-        db = self._get_db()
-        db.create_metadata_collection(name)
-
-    def is_metadata_only(self) -> bool:
-        """Check if the current collection is metadata-only.
-
-        Returns:
-            True if metadata-only, False if vector collection.
-        """
-        if self._collection is None:
-            return False
-        return self._collection.is_metadata_only()
-
-    def velesql(self, query_str: str, params: Optional[dict] = None, **kwargs: Any) -> VectorStoreQueryResult:
-        """Execute a VelesQL query."""
-        if self._collection is None:
-            return VectorStoreQueryResult(nodes=[], similarities=[], ids=[])
-        results = self._collection.query(query_str, params)
-        return self._build_query_result(results)
-
-    def multi_query_search(
-        self,
-        query_embeddings: List[List[float]],
-        similarity_top_k: int = 10,
-        fusion: str = "rrf",
-        fusion_params: Optional[dict] = None,
-        **kwargs: Any,
-    ) -> VectorStoreQueryResult:
-        """Multi-query fusion search combining results from multiple query embeddings.
-
-        Uses fusion strategies to combine results from multiple query reformulations,
-        ideal for RAG pipelines using Multiple Query Generation (MQG).
-
-        Args:
-            query_embeddings: List of query embedding vectors.
-            similarity_top_k: Number of results to return.
-            fusion: Fusion strategy ("rrf", "average", "maximum", "weighted").
-            fusion_params: Parameters for fusion strategy:
-                - RRF: {"k": 60} (default k=60)
-                - Weighted: {"avg_weight": 0.6, "max_weight": 0.3, "hit_weight": 0.1}
-            **kwargs: Additional arguments.
-
-        Returns:
-            Query result with fused nodes and scores.
-
-        Example:
-            >>> results = vector_store.multi_query_search(
-            ...     query_embeddings=[emb1, emb2, emb3],
-            ...     similarity_top_k=10,
-            ...     fusion="rrf",
-            ...     fusion_params={"k": 60}
-            ... )
-        """
-        if not query_embeddings:
-            return VectorStoreQueryResult(nodes=[], similarities=[], ids=[])
-
-        dimension = len(query_embeddings[0])
-        collection = self._get_collection(dimension)
-
-        # Build fusion strategy
-        fusion_params = fusion_params or {}
-        if fusion == "rrf":
-            k = fusion_params.get("k", 60)
-            fusion_strategy = velesdb.FusionStrategy.rrf(k=k)
-        elif fusion == "average":
-            fusion_strategy = velesdb.FusionStrategy.average()
-        elif fusion == "maximum":
-            fusion_strategy = velesdb.FusionStrategy.maximum()
-        elif fusion == "weighted":
-            avg_w = fusion_params.get("avg_weight", 0.6)
-            max_w = fusion_params.get("max_weight", 0.3)
-            hit_w = fusion_params.get("hit_weight", 0.1)
-            fusion_strategy = velesdb.FusionStrategy.weighted(
-                avg_weight=avg_w, max_weight=max_w, hit_weight=hit_w
-            )
-        else:
-            fusion_strategy = velesdb.FusionStrategy.rrf(k=60)
-
-        results = collection.multi_query_search(
-            vectors=query_embeddings,
-            top_k=similarity_top_k,
-            fusion=fusion_strategy,
-        )
-
-        return self._build_query_result(results)
-
-    def explain(
-        self,
-        query_str: str,
-        **kwargs: Any,
-    ) -> dict:
-        """Get the query execution plan for a VelesQL query."""
-        validate_query(query_str)
-        if self._collection is None:
-            raise ValueError("Collection not initialized. Add documents first.")
-        return self._collection.explain(query_str)
-
-    def match_query(
-        self,
-        query_str: str,
-        params: Optional[dict] = None,
-        **kwargs: Any,
-    ) -> VectorStoreQueryResult:
-        """Execute a MATCH query and convert results to VectorStoreQueryResult."""
-        validate_query(query_str)
-        if self._collection is None:
-            return VectorStoreQueryResult(nodes=[], similarities=[], ids=[])
-
-        results = self._collection.match_query(query_str, params=params, **kwargs)
-        nodes: List[TextNode] = []
-        similarities: List[float] = []
-        ids: List[str] = []
-
-        for result in results:
-            node_id = str(result.get("node_id", ""))
-            projected = result.get("projected", {})
-            bindings = result.get("bindings", {})
-            text = str(projected if projected else bindings)
-            score = float(result.get("score", 0.0) or 0.0)
-            nodes.append(TextNode(text=text, id_=node_id, metadata=result))
-            similarities.append(score)
-            ids.append(node_id)
-
-        return VectorStoreQueryResult(nodes=nodes, similarities=similarities, ids=ids)
-
-    def train_pq(self, m: int = 8, k: int = 256, opq: bool = False) -> str:
-        """Train Product Quantization on the collection.
-
-        PQ training is a Database-level operation (not Collection-level)
-        because TRAIN QUANTIZER requires Database-level VelesQL execution.
-
-        Args:
-            m: Number of subspaces. Defaults to 8.
-            k: Number of centroids per subspace. Defaults to 256.
-            opq: Enable Optimized PQ pre-rotation. Defaults to False.
-
-        Returns:
-            Training result message.
-        """
-        return self._get_db().train_pq(self.collection_name, m=m, k=k, opq=opq)
 
     def stream_insert(
         self,
@@ -944,24 +400,21 @@ class VelesDBVectorStore(BasePydanticVectorStore):
         if not nodes:
             return 0
 
-        # Security: Validate batch size
         validate_batch_size(len(nodes))
 
-        # Extract sparse vectors from kwargs (v1.5)
         sparse_vectors = kwargs.get("sparse_vectors")
         if sparse_vectors is not None:
             for sv in sparse_vectors:
                 validate_sparse_vector(sv)
 
-        # Get dimension from first node's embedding
         first_embedding = nodes[0].get_embedding()
         if first_embedding is None:
             raise ValueError("Nodes must have embeddings")
         dimension = len(first_embedding)
 
         collection = self._get_collection(dimension)
-
         points = []
+
         for idx, node in enumerate(nodes):
             embedding = node.get_embedding()
             if embedding is None:
@@ -971,14 +424,8 @@ class VelesDBVectorStore(BasePydanticVectorStore):
                 )
 
             node_id = node.node_id
+            payload = {"text": node.get_content(), "node_id": node_id}
 
-            # Build payload
-            payload = {
-                "text": node.get_content(),
-                "node_id": node_id,
-            }
-
-            # Add metadata
             if hasattr(node, "metadata") and node.metadata:
                 for key, value in node.metadata.items():
                     if isinstance(value, (str, int, float, bool)):
@@ -990,7 +437,6 @@ class VelesDBVectorStore(BasePydanticVectorStore):
                 "payload": payload,
             }
 
-            # Attach sparse vector if provided
             if sparse_vectors is not None and idx < len(sparse_vectors):
                 point["sparse_vector"] = sparse_vectors[idx]
 
@@ -1028,7 +474,6 @@ class VelesDBVectorStore(BasePydanticVectorStore):
         if not nodes:
             return []
 
-        # Security: Validate batch size
         validate_batch_size(len(nodes))
 
         first_embedding = nodes[0].get_embedding()
@@ -1037,7 +482,6 @@ class VelesDBVectorStore(BasePydanticVectorStore):
         dimension = len(first_embedding)
 
         collection = self._get_collection(dimension)
-
         result_ids: List[str] = []
         batch: list = []
 
@@ -1049,10 +493,7 @@ class VelesDBVectorStore(BasePydanticVectorStore):
             node_id = node.node_id
             result_ids.append(node_id)
 
-            payload = {
-                "text": node.get_content(),
-                "node_id": node_id,
-            }
+            payload = {"text": node.get_content(), "node_id": node_id}
             if hasattr(node, "metadata") and node.metadata:
                 for key, value in node.metadata.items():
                     if isinstance(value, (str, int, float, bool)):
@@ -1072,3 +513,84 @@ class VelesDBVectorStore(BasePydanticVectorStore):
             collection.stream_insert(batch)
 
         return result_ids
+
+    # ------------------------------------------------------------------
+    # Read / utility operations
+    # ------------------------------------------------------------------
+
+    def get_nodes(self, node_ids: List[str], **kwargs: Any) -> List[TextNode]:
+        """Retrieve nodes by their IDs."""
+        if not node_ids or self._collection is None:
+            return []
+        int_ids = [_stable_hash_id(nid) for nid in node_ids]
+        points = self._collection.get(int_ids)
+        result = []
+        for pt in points:
+            if pt:
+                p = pt.get("payload", {})
+                result.append(
+                    TextNode(
+                        text=p.get("text", ""),
+                        id_=p.get("node_id", ""),
+                        metadata=self._metadata_from_payload(p),
+                    )
+                )
+        return result
+
+    def get_collection_info(self) -> dict:
+        """Get collection configuration information."""
+        if self._collection is None:
+            return {
+                "name": self.collection_name,
+                "dimension": 0,
+                "metric": self.metric,
+                "point_count": 0,
+            }
+        return self._collection.info()
+
+    def flush(self) -> None:
+        """Flush all pending changes to disk."""
+        if self._collection is not None:
+            self._collection.flush()
+
+    def is_empty(self) -> bool:
+        """Check if the collection is empty."""
+        return self._collection is None or self._collection.is_empty()
+
+    def create_metadata_collection(self, name: str) -> None:
+        """Create a metadata-only collection (no vectors).
+
+        Useful for storing reference data that can be JOINed with
+        vector collections (VelesDB Premium feature).
+
+        Args:
+            name: Collection name.
+        """
+        db = self._get_db()
+        db.create_metadata_collection(name)
+
+    def is_metadata_only(self) -> bool:
+        """Check if the current collection is metadata-only.
+
+        Returns:
+            True if metadata-only, False if vector collection.
+        """
+        if self._collection is None:
+            return False
+        return self._collection.is_metadata_only()
+
+    def train_pq(self, m: int = 8, k: int = 256, opq: bool = False) -> str:
+        """Train Product Quantization on the collection.
+
+        PQ training is a Database-level operation (not Collection-level)
+        because TRAIN QUANTIZER requires Database-level VelesQL execution.
+
+        Args:
+            m: Number of subspaces. Defaults to 8.
+            k: Number of centroids per subspace. Defaults to 256.
+            opq: Enable Optimized PQ pre-rotation. Defaults to False.
+
+        Returns:
+            Training result message.
+        """
+        return self._get_db().train_pq(self.collection_name, m=m, k=k, opq=opq)
